feat: pulled image building and pushing to ecr_client
